# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GP Premium API",
        "description": "API para gerenciamento de produção de pneus recauchutados da GP Premium",
        "contact": {
            "name": "Equipe GP Premium",
            "url": "https://www.gppremium.com.br",
            "email": "contato@gppremium.com.br"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://www.gppremium.com.br/license"
        },
        "version": "0.0.35-SNAPSHOT"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor de Desenvolvimento"
        },
        {
            "url": "https://api.gppremium.com.br",
            "description": "Servidor de Produção"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
        "/api/usuario/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Consultar usuário por ID",
                "description": "Retorna os dados de um usuário específico pelo seu ID",
                "operationId": "consultar_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de autenticação inválido ou expirado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioModel"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Usuário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar usuário",
                "description": "Atualiza os dados de um usuário existente, incluindo validação de perfis",
                "operationId": "atualizarUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsuarioModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de autenticação inválido ou expirado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Excluir usuário",
                "description": "Remove um usuário do sistema",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token de autenticação inválido ou expirado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Usuário excluído com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            }
        },
        "/api/tipoobservacao/{id}": {
            "get": {
                "tags": [
                    "tipo-observacao-controller"
                ],
                "operationId": "consultar_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "tipo-observacao-controller"
                ],
                "operationId": "atualizar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoObservacaoModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tipo-observacao-controller"
                ],
                "operationId": "delete_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tipoclassificacao/{id}": {
            "get": {
                "tags": [
                    "tipo-classificacao-controller"
                ],
                "operationId": "consultar_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "tipo-classificacao-controller"
                ],
                "operationId": "atualizar_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoClassificacaoModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tipo-classificacao-controller"
                ],
                "operationId": "delete_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/regra/{id}": {
            "get": {
                "tags": [
                    "regra-controller"
                ],
                "operationId": "consultar_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "regra-controller"
                ],
                "operationId": "atualizar_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegraModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "regra-controller"
                ],
                "operationId": "delete_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/qualidade/{id}": {
            "get": {
                "tags": [
                    "Qualidade"
                ],
                "summary": "Consultar controle de qualidade por ID",
                "description": "Retorna um controle de qualidade específico pelo seu ID",
                "operationId": "consultar_6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do controle de qualidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Controle de qualidade encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Controle de qualidade não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Qualidade"
                ],
                "summary": "Atualizar controle de qualidade",
                "description": "Atualiza um controle de qualidade existente",
                "operationId": "atualizar_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do controle de qualidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QualidadeModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Controle de qualidade atualizado com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Controle de qualidade não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Qualidade"
                ],
                "summary": "Excluir controle de qualidade",
                "description": "Exclui um controle de qualidade pelo ID",
                "operationId": "delete_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do controle de qualidade",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Controle de qualidade excluído com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Controle de qualidade não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            }
        },
        "/api/producao/{id}": {
            "get": {
                "tags": [
                    "Produção"
                ],
                "summary": "Consultar produção por ID",
                "description": "Retorna os detalhes de uma produção específica pelo seu ID",
                "operationId": "consultar_7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da produção",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Token de autenticação inválido ou expirado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Produção encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProducaoDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produção não encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Produção"
                ],
                "summary": "Atualizar produção",
                "description": "Atualiza os dados de uma produção existente",
                "operationId": "atualizar_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da produção",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProducaoModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "401": {
                        "description": "Token de autenticação inválido ou expirado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produção não encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Produção atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProducaoModel"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Produção"
                ],
                "summary": "Excluir produção",
                "description": "Exclui uma produção e retorna o status da carcaça para 'start'",
                "operationId": "delete_5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da produção",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Token de autenticação inválido ou expirado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Produção não encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Produção excluída com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer Authentication": []
                    }
                ]
            }
        },
...
        "/api/espessuramento": {
            "get": {
                "tags": [
                    "espessuramento-controller"
                ],
                "operationId": "findAll_14",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EspessuramentoModel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "espessuramento-controller"
                ],
                "operationId": "salvar_12",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EspessuramentoModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UsuarioModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID único do usuário",
                        "format": "int64",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "description": "Nome completo do usuário",
                        "example": "João Silva"
                    },
                    "login": {
                        "type": "string",
                        "description": "Login único do usuário",
                        "example": "joao.silva"
                    },
                    "senha": {
                        "type": "string",
                        "description": "Senha do usuário (apenas escrita)",
                        "writeOnly": true,
                        "example": "senha123"
                    },
                    "perfil": {
                        "type": "array",
                        "description": "Lista de perfis/roles do usuário",
                        "items": {
                            "$ref": "#/components/schemas/PerfilModel"
                        }
                    }
                },
                "description": "Modelo representando um usuário do sistema"
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "description": "Token JWT para autenticação. Formato: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}